+++ packages/vite-node-miniflare/src/index.ts
@@ -1,2 +1,1 @@
-// export { vitePluginViteNodeMiniflare } from "./server/plugin";
-export { vitePluginWorkerd } from "./v6/plugin";
+export * from "./plugin";
