+++ packages/vite-node-miniflare/src/shared.ts
@@ -1,2 +1,21 @@
-export const SERVER_RPC_PATH = "/__server_rpc__";
-export const CLIENT_RPC_PATH = "/__client_rpc__";
+export const RUNNER_INIT_PATH = "/__viteInit";
+export const RUNNER_EVAL_PATH = "/__viteEval";
+export const ANY_URL = "https://any.local";
+
+export type RunnerEnv = {
+  __viteUnsafeEval: {
+    eval: (code: string, filename?: string) => any;
+  };
+  __viteFetchModule: {
+    fetch: (request: Request) => Promise<Response>;
+  };
+  __viteRunner: DurableObject;
+  __viteOptions: {
+    root: string;
+    hmr: boolean;
+  };
+};
+
+export type FetchMetadata = {
+  entry: string;
+};
