--- packages/vite-node-miniflare/src/v6/index.ts
@@ -1,1 +1,0 @@
-export * from "./plugin";
