+++ packages/vite-node-miniflare/src/worker.ts
