+++ packages/vite-node-miniflare/tsup.config.ts
@@ -2,8 +2,7 @@ import { defineConfig } from "tsup";
 
 export default [
   defineConfig({
-    // entry: ["src/client/worker-entry.ts"],
-    entry: ["src/v6/worker.ts"],
+    entry: ["src/worker.ts"],
     format: ["esm"],
     platform: "browser",
     noExternal: [/.*/],
